+++ modules/main/src/main/java/com/nisovin/shopkeepers/text/Text.java
@@ -25,7 +25,7 @@ import com.nisovin.shopkeepers.util.text.MessageArguments;
  * the {@link #getChild() child} reference, allowing them to encompass a series of {@link Text}s.
  * The next {@link Text} in the chain only inherits the formatting features from the previous
  * {@link Text}s (including the previous child {@link Text}s). The formatting can be reset by
- * inserting a {@link FormattingText} with {@link ChatColor#RESET}.
+ * inserting a {@link FormattingText} with a color code or {@link ChatColor#RESET}.
  * <p>
  * The structure and semantics of this text representation are closely oriented on plain text with
  * color codes, but still supports advanced text features such as {@link HoverEventText hover