+++ CHANGELOG.md
@@ -9,6 +9,7 @@ Date format: (YYYY-MM-DD)
   * If you want to run this plugin on an older server version, you need to use an older plugin version.
   * Internal: Various updates for Bukkit 1.20.6 and Java 21.
 * Item display name and lore data from the language file can be specified as Json text now.
+* Fix: Color codes in text (e.g. from the language file) did not reset formatting.
 * Fix: Item data containing attribute modifiers was not correctly saved to the plugin config and failed to load.
 * Extend the item stack we use for server assumption tests to include newer item data.
 * Update the compatibility mode:
@@ -19,6 +20,7 @@ Date format: (YYYY-MM-DD)
   * Use the JDT `NonNullByDefault` annotation which ECJ interprets as non-null-by-default in more contexts (e.g. for type parameters). Remove various now redundant null annotations.
   * Add `external-annotations` module for external ECJ null annotations. This reduces the need for various `assertNonNull` calls.
 * Internal: Avoid force-loading entities for chunks whose entities were not yet loaded, e.g. for the `check` command or the spawn blocking feature.
+* Internal: Avoid setting `reset` as color code.
 
 ## v2.22.3 (2024-07-27)
 ### Supported MC versions: 1.21, 1.20.6, 1.20.4, 1.20.2, 1.20.1, 1.19.4, 1.18.2, 1.17.1, 1.16.5